refactor toasts
