+++ src/betterdiscord/builtins/developer/recovery.tsx
@@ -5,9 +5,9 @@ import DiscordModules from "@modules/discordmodules";
 import {t} from "@common/i18n";
 import Builtin from "@structs/builtin";
 import Settings from "@stores/settings";
+import Toasts from "@stores/toasts";
 import pluginmanager from "@modules/pluginmanager";
 import IPC from "@modules/ipc";
-import Toasts from "@ui/toasts";
 import Modals from "@ui/modals";
 import {getByKeys, getByPrototypes, getByStrings, getMangled} from "@webpack";
