+++ src/betterdiscord/modules/addonmanager.ts
@@ -8,13 +8,13 @@ import AddonError from "@structs/addonerror";
 import Settings from "@stores/settings";
 import Events from "./emitter";
 import JsonStore, {type Files} from "@stores/json";
+import Toasts from "@stores/toasts";
 import React from "./react";
 import {t} from "@common/i18n";
 import ipc from "./ipc";
 
 import AddonEditor from "@ui/misc/addoneditor";
 import FloatingWindows from "@ui/floatingwindows";
-import Toasts from "@ui/toasts";
 import Store from "@stores/base";
 import type {SystemError} from "bun";
 import RemoteAPI from "@polyfill/remote";
