+++ src/betterdiscord/modules/thememanager.ts
@@ -1,4 +1,5 @@
 import Config from "@stores/config";
+import Toasts from "@stores/toasts";
 
 import AddonError from "@structs/addonerror";
 
@@ -6,7 +7,6 @@ import AddonManager, {type Addon} from "./addonmanager";
 import DOMManager from "./dommanager";
 import {t} from "@common/i18n";
 
-import Toasts from "@ui/toasts";
 import Modals from "@ui/modals";
